Changes; - Changed the way items are added to the Bloom filter - Fixed small issues with how the keys are hashed and checked - Added rate of false positives to script output

+++ Code/GreenOnion/Bloom/MurmurHash3.h
@@ -26,7 +26,7 @@ typedef unsigned __int64 uint64_t;
 
 //-----------------------------------------------------------------------------
 
-uint MurmurHash3_x86_32  ( const uint * key, int len, uint32_t seed);
+uint MurmurHash3_x86_32  ( const uint long * key, int len, uint32_t seed);
 
 //-----------------------------------------------------------------------------
